requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
distlib==0.4.0
filelock==3.20.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
platformdirs==4.5.0
python-dotenv==1.2.1
virtualenv==20.35.3
Werkzeug==3.1.3
